# Group async calls in bootstrapping for more concurrency