Changed the waitgroup behavior in the NFL bootstrapping to start all concurrent functions at once instead of waiting for each group to finish. I wasn't sure if there was some reason that it was done in separate waitgroups, but the calls don't appear to be dependent on each other, and the responses still seem correct at a glance.

It doesn't seem to have saved much time on the second or third bootstraps, but I think the longer DB calls in the first bootstrap (get all players, get player faces) benefit more from the concurrency.

Performance results in local testing:
- Bootstrap one:
  - Main branch: 4768ms
![image](https://github.com/user-attachments/assets/2e68ecb6-0b92-481a-8aab-be82fb0c4b17)
  - With changes: 3321ms
![image](https://github.com/user-attachments/assets/a0709e09-b4e0-4cdc-8f94-d27458e0d737)
- Bootstrap two:
  - Main branch: 1346ms
![image](https://github.com/user-attachments/assets/c20d551f-3d2a-4810-9b22-dcf78e94a608)
  - With changes: 1333ms
![image](https://github.com/user-attachments/assets/1fb632b0-1e29-4536-8b46-06018f36c3c7)
- Bootstrap three:
  - Main branch: 2782ms
![image](https://github.com/user-attachments/assets/b6ae6308-1593-4804-a062-7298fbe04c29)
  - With changes: 2668ms
![image](https://github.com/user-attachments/assets/992e39ca-ba76-4c45-a70e-4e601de30a14)